Two small objects
Neither have animations. It's worth noting that the kitchen stool and table do not go together: the "kitchen table" goes with the "kitchen chair" inbwteen the den and the kitchen, while the "kitchen stool" will go with the "kitchen island" in the actual kitchen. The kitchen island and kitchen chair are coming up soon.